Adjust CMakeLists.txt files for simpler integration
With this commit, we are able to build additional targets more easily, such as a static library. The static library is then used by the unit tests to link against. To verify the changes, we add a doctest-based unit test as a first example on how to integrate unit tests in icarus.